fix(numpy): aliases of builtin types is deprecated as of numpy 1.20

+++ autotest/test_gwf_csub_sub01_adjmat.py
@@ -270,7 +270,7 @@ def eval_sub(sim):
 
     # calculate theta and porosity from total interbed compaction
     comp = tc['TCOMP']
-    dtype = [('THICK', np.float), ('THETA', np.float)]
+    dtype = [('THICK', float), ('THETA', float)]
     ovalsi = np.zeros((comp.shape[0]), dtype=dtype)
     ovalsi['THICK'] = tc['THICK']
     ovalsi['THETA'] = tc['THETA']
@@ -382,7 +382,7 @@ def cbc_compare(sim):
             qout = 0.
             v = cobj.get_data(kstpkper=k, text=text)[0]
             if isinstance(v, np.recarray):
-                vt = np.zeros(size3d, dtype=np.float)
+                vt = np.zeros(size3d, dtype=float)
                 for jdx, node in enumerate(v['node']):
                     vt[node - 1] += v['q'][jdx]
                 v = vt.reshape(shape3d)
@@ -402,7 +402,7 @@ def cbc_compare(sim):
             key = '{}_OUT'.format(text)
             d[key][idx] = qout
 
-    diff = np.zeros((nbud, len(bud_lst)), dtype=np.float)
+    diff = np.zeros((nbud, len(bud_lst)), dtype=float)
     for idx, key in enumerate(bud_lst):
         diff[:, idx] = d0[key] - d[key]
     diffmax = np.abs(diff).max()
